+++ README.md
@@ -1,2 +1,86 @@
 # MagicMe
-team_project
+
+# 当前项目组正在做：
+
+:heart:**人脸识别打卡系统**(MagicMe):heart:
+
+:hammer:**解决问题:**
+
+结束手动签到的时代，欢迎来到刷脸打卡的新时代！！！
+
+只需往屏幕扫一眼，就能实现签到，速度快，安全高，识别准确，自动记录签到的时间和谁进行了签到。
+
+:pushpin:**打卡系统流程图：**
+
+<img src="https://typora-figure-bed.oss-cn-chengdu.aliyuncs.com/img/20210601155143.png" alt="截屏2021-06-01 下午3.51.38" style="zoom:50%;" />
+
+:rainbow_flag:**团队宗旨：**
+
+- **解放思想**：改变自己懒的思想，提高自己的生成力和技术，是解决一切问题的根本途径。
+- **实事求是**：根据实际情况，决定项目路线。
+- **齐心协力**：互相帮助，共同进步。
+- **团结一致**：遇到困难，积极交流和想办法解决，而不是*怨天*尤人。
+
+:moneybag:**捐赠(donate):**
+
+​	<img src="https://typora-figure-bed.oss-cn-chengdu.aliyuncs.com/img/20210601153340.JPG" alt="IMG_4847(20210601-153316)" style="zoom: 33%;" />
+
+# 项目方向：人脸识别
+
+- 人脸检测、识别（图片、视频）
+- 轮廓标识
+- 头像合成（给人戴帽子）
+- 数字化妆（画口红、眉毛、眼睛等）
+- 性别识别
+- 表情识别（生气、厌恶、恐惧、开心、难过、惊喜、平静等七种情绪）
+- 视频对象提取
+- 图片修复（可用于水印去除）
+- 图片自动上色
+- 眼动追踪（待完善）
+- 换脸（待完善）
+- .......
+
+注：根据具体情况和时间安排来实现几个子方向
+
+# 营销策略
+
+## 娱乐向（增量1）
+
+主要是给普通大众使用（比如手机上的APP）
+
+- 化妆
+- 换头
+- 表情识别
+- .....
+
+营销策略：免费，功能强大 ，注重用户隐私保护
+
+目的：扩大产品影响力，提高知名度
+
+## 专业向（增量2）
+
+- 比如联合公安局抓小偷
+- 安全帽识别
+- 小区门禁，公司打卡等，把人脸识别作为身份认证
+- ......
+
+营销策略：按服务收费，一年多少钱之类的。
+
+目的：为了盈利
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
