+++ UI/src/main.js
@@ -1,0 +1,18 @@
+import Vue from 'vue'
+import App from './App'
+import router from './router/index'
+import ElementUI from 'element-ui'
+import 'element-ui/lib/theme-chalk/index.css'
+import VueRouter from 'vue-router'
+
+Vue.config.productionTip = false
+
+Vue.use(ElementUI)
+Vue.use(VueRouter)
+/* eslint-disable no-new */
+new Vue({
+  el: '#app',
+  router,
+  components: { App },
+  template: '<App/>'
+})
