+++ personal log/201931072426-zy_log.md
@@ -1,0 +1,10 @@
+# 朱尧日志
+
+## 第一次
+
+本周编写了前端登录和注册模块， 使用vue进行编写，vue-axios进行交互，并编写了基于express-node.js的简易后端框架，编写了denglu.vue、zhuce.vue两个页面和app.js一个服务器。
+
+问题：
+
+使用vue进行数据交互时需使用axios而非以前使用的ajax，语法和使用有些许不同，故花费了不少时间重新学习。
+
