--- test.txt
@@ -1,7 +1,0 @@
-这是一个是测试文件，用来测试组员的pull。测试成功的人，在此文本下写下自己的名字，必须能在团队仓库看见。
-朱元凯。
-张天泳。
-裴宇航。
-王庆。
-田果。
-朱尧。
